Data type optimizations, added GameState.java

+++ PR1-K17/src/main/java/com/k17/pr1k17/GameNumber.java
@@ -10,7 +10,7 @@ public class GameNumber
         this.isSelected = false;
     }
 
-    public int getValue() {
+    public short getValue() {
         return value;
     }
